backend/requirements.txt
fastapi
uvicorn
transformers
langchain
python-dotenv
requests
pandas
PyPDF2
torch
